backend/requirements.txt
firebase-admin
spotipy
python-dotenv